+++ src/main/java/subway/domain/repository/StationRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
